Refactor camera & culling to render module
