+++ src/world/ecs/mod.rs
@@ -4,7 +4,6 @@ use crate::vector3;
 use cgmath::{One, Quaternion, Vector3};
 
 pub mod bounds;
-pub mod camera;
 pub mod chunk_loader;
 pub mod physics;
 pub mod player;
